+++ zcash_primitives/CHANGELOG.md
@@ -21,9 +21,13 @@ and this library adheres to Rust's notion of
   methods for these types that do not exist for the `incrementalmerkletree`
   replacement types have been replaced by new methods in the `merkle_tree` module.
 - `merkle_tree::incremental::write_auth_fragment_v1` has been removed without replacement.
-- The following have been removed from `merkle_tree::incremental`; these were
-  `zcashd`-specific serialization methods which have been moved into the
+- The `merkle_tree::incremental` module has been removed; its former contents
+  were either moved to the `merkle_tree` module or were `zcashd`-specific
+  serialization methods which have been removed entirely and moved into the
   [zcashd](https://github.com/zcash/zcash) repository.
+- The dependency on the `bridgetree` crate has been removed from
+  `zcash_primitives` and the following zcashd-specific serialization methods
+  have been moved to the [zcashd](https://github.com/zcash/zcash) repository:
   - `read_auth_fragment_v1`
   - `read_bridge_v1`
   - `read_bridge_v2`
@@ -34,6 +38,24 @@ and this library adheres to Rust's notion of
   - `write_checkpoint_v2`
   - `read_tree`
   - `write_tree`
+
+### Moved
+- The following constants and methods have been moved from the
+  `merkle_tree::incremental` module into the `merkle_tree` module to
+  consolidate the serialization code for commitment tree frontiers:
+  - `SER_V1`
+  - `SER_V2`
+  - `write_usize_leu64`
+  - `read_leu64_usize`
+  - `write_position`
+  - `read_position`
+  - `write_address`
+  - `read_address`
+  - `read_frontier_v0`
+  - `write_nonempty_frontier`
+  - `read_nonempty_frontier_v1`
+  - `write_frontier_v1`
+  - `read_frontier_v1`
 
 ### Added
 - `merkle_tree::incremental::{read_address, write_address}`
